+++ src/Resources/EntityDataResource.php
@@ -40,6 +40,16 @@ class EntityDataResource extends Resource
 
     protected static ?string $slug = 'ff-data';
 
+    public static function isScopedToTenant(): bool
+    {
+        return config('flex-fields.tenancy.enabled', false) && config('flex-fields.tenancy.is_tenant_aware.records', true);
+    }
+
+    public static function getTenantOwnershipRelationshipName(): string
+    {
+        return 'tenant';
+    }
+
     public static function form(Schema $schema): Schema
     {
         $entity = static::getCurrentEntity();